backend/requirements.txt
Django==4.1.2
djangorestframework==3.14.0
django-rest-swagger==2.2.0
psycopg2==2.9.4
python-dotenv==0.21.0